requirements.txt
Flask
flask_cors
requests
google-cloud-storage
google-cloud-datastore